+++ pulsar/async/eventloop.py
@@ -26,7 +26,7 @@ from .udp import create_datagram_endpoint
 from .consts import DEFAULT_CONNECT_TIMEOUT, DEFAULT_ACCEPT_TIMEOUT
 from .pollers import DefaultIO
 
-__all__ = ['EventLoop', 'TimedCall']
+__all__ = ['EventLoop', 'TimedCall', 'run_in_loop_thread']
 
 
 def file_descriptor(fd):
@@ -41,6 +41,23 @@ def setid(self):
     self.tid = ct.ident
     self.pid = os.getpid()
     return ct
+
+
+def run_in_loop_thread(event_loop, callback, *args, **kwargs):
+    '''Run ``callable`` in the ``event_loop`` thread.
+
+    Return a :class:`Deferred`
+    '''
+    d = Deferred()
+
+    def _():
+        try:
+            result = yield callback(*args, **kwargs)
+        except Exception:
+            result = sys.exc_info()
+        d.set_result(result)
+    event_loop.call_soon_threadsafe(_)
+    return d
 
 
 class EventLoopPolicy(BaseEventLoopPolicy):